comments and calc changes
added some comments and changed calculation of distance. Also changed sending of diameter value

+++ nRF5340_code/src/dataCSC.h
@@ -52,7 +52,7 @@ public:
     uint8_t type;
 
     // received data from application
-    uint8_t wheelDiameter;
+    double wheelDiameter;
 
     // attributes for calculations
 private:
